refact: move RenderEffect to timeline package
